Creamos la clase LectorCSV y seguimos con rondas
la clase tiene todo lo que estaba en el main, ahora el main esta totalmente limpio. esto nos permite meter mas facilmente la base de datos.
