+++ Entrega1/src/Participante.java
@@ -26,8 +26,11 @@ public class Participante {
     public String getNombre() {
         return nombre;
     }
-    public void sumaPuntosFinales() {
-       puntosFinales += puntos;
+    public int getPuntosFinales() {
+        return puntosFinales;
+    }
+    public void setPuntosFinales(int resultadoRondaFinal) {
+        this.puntosFinales = resultadoRondaFinal;
     }
 
 }
